requirements.txt
flask==2.2.3
pymongo==4.3.3
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
certifi==2022.12.7
matplotlib==3.7.1
python-dotenv==1.0.0
apscheduler==3.10.1
gunicorn==20.1.0